quiz-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
dotenv==0.9.9
ecdsa==0.19.1
elevenlabs==2.26.1
executing==2.2.1
fastapi==0.135.2
fastjsonschema==2.21.2
fqdn==1.5.1
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
Js2Py==0.74
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-text-splitters==1.0.0
langsmith==0.4.53
lark==1.3.1
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mistune==3.2.0
MouseInfo==0.1.3
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.3.4
openai==2.8.1
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pillow==12.0.0
pipwin==0.5.2
platformdirs==4.5.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
pure_eval==0.2.3
pyasn1==0.6.3
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.19.2
pyjsparser==2.7.1
pymongo==4.16.0
PyMsgBox==2.0.1
pyperclip==1.11.0
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==4.0.0
pytweening==1.2.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
Send2Trash==2.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==1.0.0
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.1
uuid_utils==0.12.0
uvicorn==0.42.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
yarl==1.22.0
zstandard==0.25.0
